+++ src/utils.ts
@@ -256,10 +256,13 @@ export function userOptionsToOptions(defaultOptions: Options, userOptions: UserO
   }
 }
 
-// https://github.com/puppeteer/puppeteer/blob/master/docs/api.md#requestrespondresponse
 export const getBody = (body: Object | string | undefined | null): string | undefined => {
-  if (typeof body === 'string' || typeof body === 'undefined') {
+  if (typeof body === 'string') {
     return body
+  }
+
+  if (body == null) {
+    return undefined
   }
 
   return JSON.stringify(body)